# Holds function types as values, not ptrs in wasm.Module