--- a/internal/engine/compiler/impl_amd64_test.go
+++ b/internal/engine/compiler/impl_amd64_test.go
@@ -49,7 +49,7 @@ func TestAmd64Compiler_indirectCallWithTargetOnCallingConvReg(t *testing.T) {
 
 	compiler := env.requireNewCompiler(t, newCompiler, &wazeroir.CompilationResult{
 		Signature: &wasm.FunctionType{},
-		Types:     []*wasm.FunctionType{{}},
+		Types:     []wasm.FunctionType{{}},
 		HasTable:  true,
 	}).(*amd64Compiler)
 	err := compiler.compilePreamble()