--- a/internal/engine/compiler/impl_arm64_test.go
+++ b/internal/engine/compiler/impl_arm64_test.go
@@ -46,7 +46,7 @@ func TestArm64Compiler_indirectCallWithTargetOnCallingConvReg(t *testing.T) {
 
 	compiler := env.requireNewCompiler(t, newCompiler, &wazeroir.CompilationResult{
 		Signature: &wasm.FunctionType{},
-		Types:     []*wasm.FunctionType{{}},
+		Types:     []wasm.FunctionType{{}},
 		HasTable:  true,
 	}).(*arm64Compiler)
 	err := compiler.compilePreamble()