--- a/internal/testing/binaryencoding/section.go
+++ b/internal/testing/binaryencoding/section.go
@@ -16,9 +16,10 @@ func encodeSection(sectionID wasm.SectionID, contents []byte) []byte {
 //
 // See EncodeFunctionType
 // See https://www.w3.org/TR/2019/REC-wasm-core-1-20191205/#type-section%E2%91%A0
-func encodeTypeSection(types []*wasm.FunctionType) []byte {
+func encodeTypeSection(types []wasm.FunctionType) []byte {
 	contents := leb128.EncodeUint32(uint32(len(types)))
-	for _, t := range types {
+	for i := range types {
+		t := &types[i]
 		contents = append(contents, EncodeFunctionType(t)...)
 	}
 	return encodeSection(wasm.SectionIDType, contents)