--- a/internal/wasm/binary/function.go
+++ b/internal/wasm/binary/function.go
@@ -9,50 +9,48 @@ import (
 	"github.com/tetratelabs/wazero/internal/wasm"
 )
 
-func decodeFunctionType(enabledFeatures api.CoreFeatures, r *bytes.Reader) (*wasm.FunctionType, error) {
+func decodeFunctionType(enabledFeatures api.CoreFeatures, r *bytes.Reader, ret *wasm.FunctionType) (err error) {
 	b, err := r.ReadByte()
 	if err != nil {
-		return nil, fmt.Errorf("read leading byte: %w", err)
+		return fmt.Errorf("read leading byte: %w", err)
 	}
 
 	if b != 0x60 {
-		return nil, fmt.Errorf("%w: %#x != 0x60", ErrInvalidByte, b)
+		return fmt.Errorf("%w: %#x != 0x60", ErrInvalidByte, b)
 	}
 
 	paramCount, _, err := leb128.DecodeUint32(r)
 	if err != nil {
-		return nil, fmt.Errorf("could not read parameter count: %w", err)
+		return fmt.Errorf("could not read parameter count: %w", err)
 	}
 
 	paramTypes, err := decodeValueTypes(r, paramCount)
 	if err != nil {
-		return nil, fmt.Errorf("could not read parameter types: %w", err)
+		return fmt.Errorf("could not read parameter types: %w", err)
 	}
 
 	resultCount, _, err := leb128.DecodeUint32(r)
 	if err != nil {
-		return nil, fmt.Errorf("could not read result count: %w", err)
+		return fmt.Errorf("could not read result count: %w", err)
 	}
 
 	// Guard >1.0 feature multi-value
 	if resultCount > 1 {
 		if err = enabledFeatures.RequireEnabled(api.CoreFeatureMultiValue); err != nil {
-			return nil, fmt.Errorf("multiple result types invalid as %v", err)
+			return fmt.Errorf("multiple result types invalid as %v", err)
 		}
 	}
 
 	resultTypes, err := decodeValueTypes(r, resultCount)
 	if err != nil {
-		return nil, fmt.Errorf("could not read result types: %w", err)
+		return fmt.Errorf("could not read result types: %w", err)
 	}
 
-	ret := &wasm.FunctionType{
-		Params:  paramTypes,
-		Results: resultTypes,
-	}
+	ret.Params = paramTypes
+	ret.Results = resultTypes
 
 	// cache the key for the function type
 	_ = ret.String()
 
-	return ret, nil
+	return nil
 }