… experimental/logging/log_listener_test.go
… internal/engine/compiler/compiler_controlflow_test.go
… internal/integration_test/bench/hostfunc_bench_test.go
… internal/integration_test/engine/adhoc_test.go
… internal/testing/binaryencoding/encoder_test.go
… internal/wasm/binary/decoder_test.go
… internal/wasm/binary/function_test.go
… internal/wasm/counts_test.go
… internal/wasm/func_validation_test.go
… internal/wasm/function_definition_test.go
… internal/wasm/module_test.go
… internal/wasm/store_test.go
… internal/wasm/table_test.go
… internal/wazeroir/compiler_test.go
… internal/wazeroir/signature_test.go
… runtime_test.go