```
$ benchstat old.txt new.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-32             35.6µs ± 2%    35.9µs ± 3%  +0.81%  (p=0.028 n=26+26)
Initialization/interpreter-multiple-32    39.3µs ±11%    38.2µs ±10%  -2.80%  (p=0.030 n=30+30)
Initialization/compiler-32                28.9µs ± 8%    28.2µs ±11%  -2.61%  (p=0.038 n=30+30)
Initialization/compiler-multiple-32       24.5µs ± 8%    24.2µs ±10%    ~     (p=0.374 n=30+30)
Compilation/with_extern_cache-32           205µs ± 1%     201µs ± 1%  -1.96%  (p=0.000 n=30+29)
Compilation/without_extern_cache-32       5.89ms ± 1%    5.85ms ± 1%  -0.58%  (p=0.000 n=30+30)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-32              136kB ± 0%     136kB ± 0%    ~     (p=0.941 n=29+29)
Initialization/interpreter-multiple-32     137kB ± 0%     137kB ± 0%  +0.00%  (p=0.000 n=25+26)
Initialization/compiler-32                 137kB ± 0%     137kB ± 0%    ~     (p=0.292 n=30+30)
Initialization/compiler-multiple-32        142kB ± 0%     142kB ± 0%    ~     (p=0.860 n=28+28)
Compilation/with_extern_cache-32          55.2kB ± 0%    55.2kB ± 0%  -0.07%  (p=0.000 n=29+29)
Compilation/without_extern_cache-32       1.99MB ± 0%    1.99MB ± 0%    ~     (p=0.129 n=29+30)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-32               38.0 ± 0%      38.0 ± 0%    ~     (all equal)
Initialization/interpreter-multiple-32      57.0 ± 0%      57.0 ± 0%    ~     (all equal)
Initialization/compiler-32                  38.0 ± 0%      38.0 ± 0%    ~     (all equal)
Initialization/compiler-multiple-32         47.0 ± 0%      47.0 ± 0%    ~     (all equal)
Compilation/with_extern_cache-32           1.04k ± 0%     1.02k ± 0%  -1.18%  (p=0.000 n=29+30)
Compilation/without_extern_cache-32        32.7k ± 0%     32.7k ± 0%  -0.04%  (p=0.000 n=29+30)
```